Update Java versions in README
The default version of Java was already at 17 because that is the first version supported by this buildpack. This PR just updates the list of Java versions mentioned in the README to include versions that are shipped with the buildpack. It just lists 17 because this is just an example, not an exhaustive list and 17 is a LTS version so it will be around for a while.

Signed-off-by: Daniel Mikusa <dan@mikusa.com>